Create Order.java
The Order class is used to calculate the price of the desk orders based on multiple variables. The Order class has fields for the quantity, and it has a Desk object field. This class also shows the use of a static field to count how many times an Order object was instantiated.